feat: it creates pagination component and reduces dashboard complexity
