+++ src/components/index.ts
@@ -8,6 +8,7 @@ export { default as Icon } from './Icon'
 export { default as Loader } from './Loader'
 export { default as Logo } from './Logo'
 export { default as MedicationsDataTableRows } from './MedicationsDataTableRows'
+export { default as Pagination } from './Pagination'
 export { default as Table } from './Table'
 export { default as TableRow } from './TableRow'
 export { default as Text } from './Text'
